vps_server/requirements.txt
pyyaml>=6.0
requests>=2.31.0
netCDF4>=1.6
numpy>=1.24
